+++ R/compounds-set.R
@@ -40,6 +40,8 @@ setMethod("updateSetConsensus", "compoundsSet", function(obj)
     return(obj)
 })
 
+setMethod("getFragInfo", "compoundsSet", doFeatAnnGetFragInfoSets)
+
 setMethod("mergedConsensusNames", "compoundsSet", doFeatAnnMCNSets)
 setMethod("mergedConsensusNames", "compoundsConsensusSet", doFeatAnnMCNSetsCons)
 
@@ -293,7 +295,6 @@ generateCompoundsSet <- function(fGroupsSet, MSPeakListsSet, adduct, generator, 
     
     setObjects <- Map(unsetFGroupsList, msplArgs, setArgs,
                       f = function(fg, mspl, sa) do.call(generator, c(list(fGroups = fg, MSPeakLists = mspl[[1]], adduct = NULL, ...), sa)))
-    setObjects <- initSetFragInfos(setObjects, MSPeakListsSet)
 
     cons <- makeFeatAnnSetConsensus(setObjects, names(fGroupsSet), setThreshold, setThresholdAnn, setAvgSpecificScores,
                                     NULL)